requirements.txt
python-dotenv
websocket-client
kafka-python
redis
streamlit